Bugfixing done, fast versions working
- Now all 4 variants are supported (fast and standard, simple and ffnn sim measure)
- The fast versions compute the embedding /  encoding of all examples only ones at startup and also the one of the currently tested example only once instead of during each comparison with another example. 
- Classifications results by class
- Live classifications should be working with all variants

+++ data_processing/CaseBaseExtraction.py
@@ -62,8 +62,8 @@ def main():
         casebase_features_list.extend(x_train[new_indcies[i]])
 
     # transform list of values back into an array and safe to file
-    casebase_labels = np.stack(casebase_labels_list, axis=0)
-    casebase_features = np.stack(casebase_features_list, axis=0)
+    casebase_labels = np.stack(casebase_labels_list, axis=0).astype('float32')
+    casebase_features = np.stack(casebase_features_list, axis=0).astype('float32')
 
     print('Number of exaples in training data set:', casebase_features.shape[0])
